# ExplorerFrontend/app/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZondScan Public API",
    "version": "1.1.0",
...
  },
  "servers": [{ "url": "https://zondscan.com" }],
...
  "paths": {
...
    "/api/qns/resolve/{name}": {
      "get": {
        "tags": ["Addresses"],
        "operationId": "resolveQnsName",
        "summary": "Resolve a QNS name",
        "description": "Resolves a conservatively normalized ASCII name below .qrl to its canonical checksummed QIP-55 address. Resolution is enabled only when the backend has a registry address and expected chain ID configured, and it verifies both the connected chain and registry bytecode before reading records. Successful deployment checks and name resolutions are cached server side for 15 seconds. Rate limited per IP: bucket of 30 requests, refilling 30 per minute.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "QNS name below .qrl. ASCII uppercase folds to lowercase; labels accept a-z, 0-9, and hyphen. Empty labels, Unicode, and the reserved double-hyphen pattern in positions three and four are rejected.",
            "schema": { "type": "string", "minLength": 5, "maxLength": 255 },
            "example": "moscowchill.qrl"
          }
        ],
        "responses": {
          "200": {
            "description": "The normalized name and its canonical QIP-55 address.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string", "description": "Normalized lowercase QNS name." },
                    "address": { "type": "string", "description": "Canonical uppercase-Q, SHAKE256-checksummed QIP-55 address." }
                  },
                  "required": ["name", "address"]
                },
                "example": {
                  "name": "moscowchill.qrl",
                  "address": "QaaaAAaaaAAaAaaAaAAAAaAAAaAaAaAAaAaaAaaaaAAAAAAAAaAAAAaAaAAaaAAaaaaAaAAAAaaAaAAaaaaaaAaAAaaaaAaAaaaaAaaaAAAAaAAAAaAaAaaAAAaAaaaAA"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported QNS name.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "invalid QNS name" } } }
          },
          "404": {
            "description": "The name has no resolver or its resolver has no address record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "enum": ["QNS name has no resolver", "QNS name has no address record"] },
                    "name": { "type": "string", "description": "Normalized QNS name." }
                  },
                  "required": ["error", "name"]
                },
                "example": { "error": "QNS name has no address record", "name": "moscowchill.qrl" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "502": {
            "description": "The node RPC failed or returned a malformed QNS response.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "QNS resolution failed" } } }
          },
          "503": {
            "description": "QNS is unconfigured, the configured chain does not match, or registry bytecode is unavailable.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": {
                  "unconfigured": { "value": { "error": "QNS resolution is not configured" } },
                  "chainMismatch": { "value": { "error": "QNS deployment does not match the connected chain" } },
                  "registryUnavailable": { "value": { "error": "QNS registry is unavailable on the connected chain" } }
                }
              }
            }
          },
          "504": {
            "description": "The bounded QNS lookup timed out.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "QNS resolution timed out" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "InternalError": {
        "description": "Internal server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Per-IP rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "rate limit exceeded" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
    }
  }
}